creates TestHomeSection
