+++ tests/functional_test/app/src/test/java/PageObjects/HomeSection.java
@@ -7,12 +7,10 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class HomeSection {
 
-	static WebDriver driver;
-	static WebDriverWait wait;
+	WebDriver driver;
+	WebDriverWait wait;
 
-	static By introText = By.xpath("//h1[@class='vc_custom_heading vc_custom_1542126455341']");
-	By userNameTextBox = By.id("email");
-	By passwordTextBox = By.id("pass");
+	By introText = By.xpath("//h1[@class='vc_custom_heading vc_custom_1542126455341']");
 
 	/**
 	 * parameterized constructor to initialize instance variables
@@ -20,30 +18,15 @@ public class HomeSection {
 	 * @param driver browser driver of type WebDriver interface
 	 */
 	public HomeSection(WebDriver driver){
-		HomeSection.driver = driver;
+		this.driver = driver;
 		wait = new WebDriverWait(driver, 20);
-	}
-
-	public void setUserName(String strUserName){
-		driver.findElement(userNameTextBox).sendKeys(strUserName);
-	}
-
-	public void setPassword(String strPassword){
-		driver.findElement(passwordTextBox).sendKeys(strPassword);
-	}
-
-	/**
-	 * Clicks on login button
-	 */
-	public void clickLoginButton(){
-		driver.findElement(loginButton).click();
 	}
 
 	/**
 	 * Gets intro text
 	 * @return String
 	 */
-	public static  String getIntroText(){
+	public String getIntroText(){
 		wait.until(webDriver -> ExpectedConditions
 			.visibilityOfElementLocated(introText).apply(webDriver));
 
@@ -56,13 +39,13 @@ public class HomeSection {
 	 * @param strPassword account password
 	 * @return String
 	 */
-	public String loginToFacebook(String fullname, String email, String msg){
+	/*public String loginToFacebook(String fullname, String email, String msg){
 
 		this.setUserName(strUserName);
 
 		this.setPassword(strPassword);
 
 		this.clickLoginButton();
-	}
+	}*/
 
 }
